Use amino in baseapp tests
